requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
elasticsearch==8.12.0
pymongo==4.6.1
pydantic==2.6.1
pydantic-settings==2.1.0
sentence-transformers==2.3.1
faiss-cpu
pandas==2.2.0
numpy==1.26.4
python-multipart==0.0.9
httpx==0.26.0
